requirements.txt
flask>=3.0.0
llama-cpp-python>=0.2.56
sentence-transformers>=2.7.0
numpy>=1.26.0
openpyxl>=3.1.0